requirements.txt
asgiref==3.7.2
Django==4.2.4
flake8==6.1.0
isort==5.12.0
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
sqlparse==0.4.4
typing_extensions==4.7.1

dj-config-url~=0.1.1
pip~=23.2.1
Pillow~=9.5.0
setuptools~=68.0.0
numpy~=1.25.0